dto added

+++ app/src/main/java/com/example/playlistmaker/data/model/TrackDto.kt
@@ -7,7 +7,7 @@ data class TrackDto(
     val trackName: String, // Название композиции, не отображаем трек в списке, если нет поля
     val artistName: String?, // Имя исполнителя
     @SerializedName("trackTimeMillis")
-    val trackTime: Int?, // Продолжительность трека
+    val trackTime: Long?, // Продолжительность трека
     @SerializedName("artworkUrl100")
     val imageUrl: String?, // Ссылка на изображение обложки
     val collectionName: String?, // Название альбома
